requirements.txt
pandas>=2.0.0
pulp>=2.7.0
pyyaml>=6.0
tabulate>=0.9.0
python-dateutil>=2.8.0
pytest>=7.0.0
